refactor: Move matrix function to `matrix.ts`
