+++ components.json
@@ -12,6 +12,6 @@
   },
   "aliases": {
     "components": "@/components",
-    "utils": " @/utils"
+    "utils": "@/utils"
   }
 }
