+++ src/containers/NavigationSection/hooks/useUserData.ts
@@ -1,6 +1,7 @@
 import useMatrixClient from "@/hooks/matrix/useMatrixClient"
 import useLocalStorage, {LocalStorageKey} from "@/hooks/util/useLocalStorage"
-import {cleanDisplayName, type Credentials, getImageUrl} from "@/utils/util"
+import {getImageUrl} from "@/utils/matrix"
+import {cleanDisplayName, type Credentials} from "@/utils/util"
 import {useCallback, useEffect, useState} from "react"
 
 export enum UserDataState {
