+++ src/containers/RoomContainer/RoomInvitedSplash.tsx
@@ -3,7 +3,7 @@ import Typography, {TypographyVariant} from "@/components/Typography"
 import {useMemo, useState, type FC} from "react"
 import RoomNotFoundSplash from "./RoomNotFoundSplash"
 import useActiveRoomIdStore from "@/hooks/matrix/useActiveRoomIdStore"
-import {getImageUrl} from "@/utils/util"
+import {getImageUrl} from "@/utils/matrix"
 import useMatrixClient from "@/hooks/matrix/useMatrixClient"
 import useTranslation from "@/hooks/util/useTranslation"
 import {LangKey} from "@/lang/allKeys"
