+++ src/containers/RoomContainer/hooks/useRoomChat.tsx
@@ -10,7 +10,7 @@ import {type VideoMessageData} from "@/components/VideoMessage"
 import useActiveRoomIdStore from "@/hooks/matrix/useActiveRoomIdStore"
 import useEventListener from "@/hooks/matrix/useEventListener"
 import useMatrixClient from "@/hooks/matrix/useMatrixClient"
-import {getImageUrl, sendAudioMessage} from "@/utils/util"
+import {getImageUrl, sendAudioMessage} from "@/utils/matrix"
 import {MsgType, type Room, RoomMemberEvent} from "matrix-js-sdk"
 import {useEffect, useState} from "react"
 import {type MessageSendRequest} from "../ChatInput"
