+++ src/utils/util.ts
@@ -1,8 +1,5 @@
 import {getEmojiByIndex} from "@/hooks/util/useEmojiPicker"
 import dayjs from "dayjs"
-import {EventType, MsgType, type MatrixClient} from "matrix-js-sdk"
-import {type RoomMessageEventContent} from "matrix-js-sdk/lib/types"
-import {mxcUrlToHttp} from "./matrix"
 
 export enum ViewPath {
   App = "/",
@@ -34,7 +31,6 @@ export enum CommonAssertion {
   UserIdNotFound = "The client should be logged in.",
   EventSenderNotFount = "The event should has a send origin",
   AvatarUrlNotValid = "The avatar url should be valid.",
-  MessageIdEmpty = "The message id should not be empty.",
 }
 
 export function assert(
@@ -60,36 +56,6 @@ export function validateUrl(url: string): boolean {
   } catch {
     return false
   }
-}
-
-const mxcRegex = /^mxc:\/\/([^/]+)\/([^/]+)$/
-
-export function validateMxcUrl(url: string): boolean {
-  const match = url.match(mxcRegex)
-
-  if (match === null) {
-    throw new Error(
-      "Invalid MXC URL format. Must be mxc://<server-name>/<media-id>"
-    )
-  }
-
-  const [_, serverName, mediaId] = match
-
-  if (!serverName || serverName.length === 0) {
-    throw new Error("Server name cannot be empty")
-  }
-
-  if (!mediaId || mediaId.length === 0) {
-    throw new Error("Media ID cannot be empty")
-  }
-
-  return true
-}
-
-const matrixUserRegex = new RegExp(/^@[\w+.-]+:matrix\.org$/)
-
-export function validateMatrixUser(input: string): boolean {
-  return matrixUserRegex.test(input)
 }
 
 function stringToIndex(str: string): number {
@@ -161,120 +127,6 @@ export function strCapitalize(str: string): string {
 }
 
 // #region Matrix SDK utils
-export function getImageUrl(
-  url: string | null | undefined,
-  client: MatrixClient | null,
-  size?: number
-): string | undefined {
-  if (
-    url === null ||
-    url === undefined ||
-    client === null ||
-    url.length === 0
-  ) {
-    return undefined
-  }
-
-  const imageUrl =
-    size === undefined
-      ? mxcUrlToHttp(client, url, true)
-      : mxcUrlToHttp(client, url, true, size, size, "scale")
-
-  return imageUrl ?? undefined
-}
-
-export function getFileUrl(
-  url: string | null | undefined,
-  client: MatrixClient | null
-): string | undefined {
-  if (
-    url === null ||
-    url === undefined ||
-    client === null ||
-    url.length === 0
-  ) {
-    return
-  }
-
-  return client.mxcUrlToHttp(url) ?? undefined
-}
-
-export type ImageUploadedInfo = {
-  matrixUrl: string
-  filename: string
-  info: {
-    w: number
-    h: number
-    mimetype: string
-    size: number
-  }
-}
-
-export async function sendImageMessageFromFile(
-  image: File,
-  client: MatrixClient | null,
-  destinationRoom: string | null
-): Promise<void> {
-  if (client === null || destinationRoom === null) {
-    return
-  }
-
-  const imageUploadedInfo = await uploadImageToMatrix(
-    image,
-    client,
-    _percent => {}
-  )
-
-  if (imageUploadedInfo === null) {
-    return
-  }
-
-  await client.sendImageMessage(
-    destinationRoom,
-    imageUploadedInfo.matrixUrl,
-    {...imageUploadedInfo.info},
-    imageUploadedInfo.filename
-  )
-}
-
-export async function uploadImageToMatrix(
-  file: File,
-  client: MatrixClient,
-  progressCallback: (percent: number) => void
-): Promise<ImageUploadedInfo | null> {
-  try {
-    const uploadResponse = await client.uploadContent(file, {
-      type: file.type,
-      progressHandler: progress => {
-        progressCallback(Math.round((progress.loaded / progress.total) * 100))
-      },
-    })
-
-    const imageUrl = URL.createObjectURL(file)
-    const image = await getImage(imageUrl)
-
-    const height = image.height
-    const width = image.width
-
-    URL.revokeObjectURL(imageUrl)
-
-    return {
-      matrixUrl: uploadResponse.content_uri,
-      filename: file.name,
-      info: {
-        w: width,
-        h: height,
-        mimetype: file.type,
-        size: file.size,
-      },
-    }
-  } catch (error) {
-    console.error("Error uploading file:", error)
-  }
-
-  return null
-}
-
 export async function getImage(data: string): Promise<HTMLImageElement> {
   assert(isValidObjectURL(data), "Invalid object URL")
 
@@ -289,30 +141,6 @@ export async function getImage(data: string): Promise<HTMLImageElement> {
   })
 }
 
-export async function sendAudioMessage(
-  audioBlob: Blob,
-  client: MatrixClient | null,
-  destinationRoom: string | null
-): Promise<void> {
-  if (client === null || destinationRoom === null) {
-    return
-  }
-
-  const uploadResponse = await client.uploadContent(audioBlob, {
-    type: audioBlob.type,
-  })
-
-  await client.sendMessage(destinationRoom, {
-    msgtype: MsgType.Audio,
-    body: "Audio message",
-    url: uploadResponse.content_uri,
-    info: {
-      mimetype: "audio/ogg",
-      size: audioBlob.size,
-    },
-  })
-}
-
 function isValidObjectURL(url: string): boolean {
   try {
     const objUrl = new URL(url)
@@ -321,26 +149,6 @@ function isValidObjectURL(url: string): boolean {
   } catch {
     return false
   }
-}
-
-export function deleteMessage(
-  client: MatrixClient,
-  roomId: string,
-  eventId: string
-): void {
-  client.redactEvent(roomId, eventId).catch(error => {
-    console.error("Error deleting message", error)
-  })
-}
-
-export function getUsernameByUserId(userId: string): string {
-  return userId.replace(":matrix.org", "")
-}
-
-export const generateUniqueNumber = (): number => {
-  const timestamp = Date.now()
-  const randomNum = Math.floor(Math.random() * 100_000)
-  return Number.parseInt(`${timestamp}${randomNum}`)
 }
 
 export function generateRandomId(length: number = 10): string {
@@ -358,121 +166,4 @@ export function generateRandomId(length: number = 10): string {
 
 export async function delay(ms: number): Promise<void> {
   await new Promise(resolve => setTimeout(resolve, ms))
-}
-
-export async function sendVideoMessageFromFile(
-  videoFile: File,
-  client: MatrixClient | null,
-  destinationRoom: string | null,
-  onProgressHandler: (progress: number) => void
-): Promise<void> {
-  if (client === null || destinationRoom === null) {
-    return
-  }
-
-  try {
-    const uploadResponse = await client.uploadContent(videoFile, {
-      name: videoFile.name,
-      includeFilename: true,
-      type: videoFile.type,
-      progressHandler: progress => {
-        onProgressHandler(Math.round((progress.loaded / progress.total) * 100))
-      },
-    })
-
-    if (uploadResponse.content_uri !== undefined) {
-      const content: RoomMessageEventContent = {
-        body: videoFile.name,
-        info: {
-          mimetype: videoFile.type,
-          size: videoFile.size,
-        },
-        msgtype: MsgType.Video,
-        url: uploadResponse.content_uri,
-      }
-
-      void client.sendEvent(destinationRoom, EventType.RoomMessage, content)
-    }
-  } catch (error) {
-    console.error(error)
-  }
-}
-
-export async function sendFileMessageFromFile(
-  file: File,
-  client: MatrixClient | null,
-  destinationRoom: string | null,
-  onProgressHandler: (progressUpload: number) => void
-): Promise<void> {
-  if (client === null || destinationRoom === null) {
-    return
-  }
-
-  try {
-    const uploadResponse = await client.uploadContent(file, {
-      name: file.name,
-      includeFilename: true,
-      type: file.type,
-      progressHandler(progress) {
-        onProgressHandler(Math.round(progress.loaded / progress.total) * 100)
-      },
-    })
-
-    if (uploadResponse.content_uri !== undefined) {
-      const content: RoomMessageEventContent = {
-        body: file.name,
-        filename: file.name,
-        info: {
-          mimetype: file.type,
-          size: file.size,
-        },
-        msgtype: MsgType.File,
-        url: uploadResponse.content_uri,
-      }
-
-      void client.sendEvent(destinationRoom, EventType.RoomMessage, content)
-    }
-  } catch (error) {
-    console.error(error)
-  }
-}
-
-export async function sendAudioMessageFromFile(
-  file: File,
-  client: MatrixClient | null,
-  destinationRoom: string | null,
-  duration: number,
-  onProgressHandler: (progressUpload: number) => void
-): Promise<void> {
-  if (client === null || destinationRoom === null) {
-    return
-  }
-
-  try {
-    const uploadResponse = await client.uploadContent(file, {
-      name: file.name,
-      includeFilename: false,
-      type: file.type,
-      progressHandler(progress) {
-        onProgressHandler(Math.round(progress.loaded / progress.total) * 100)
-      },
-    })
-
-    if (uploadResponse.content_uri !== undefined) {
-      const content: RoomMessageEventContent = {
-        body: file.name,
-        info: {
-          duration,
-          mimetype: file.type,
-          size: file.size,
-        },
-        msgtype: MsgType.Audio,
-        url: uploadResponse.content_uri,
-      }
-
-      void client.sendEvent(destinationRoom, EventType.RoomMessage, content)
-    }
-  } catch (error) {
-    console.error(error)
-  }
 }
